+++ apps/designer-standalone/src/components/contextSettings.tsx
@@ -1,11 +1,11 @@
 import type { RootState, AppDispatch } from '../state/store';
-import { loadRun, loadWorkflow, setDarkMode, setMonitoringView, setReadOnly } from '../state/workflowLoadingSlice';
+import { setConsumption, setDarkMode, setMonitoringView, setReadOnly, loadRun, loadWorkflow } from '../state/workflowLoadingSlice';
 import { Checkbox } from '@fluentui/react';
 import { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 const ContextSettings = () => {
-  const { readOnly, monitoringView, darkMode } = useSelector((state: RootState) => {
+  const { readOnly, monitoringView, darkMode, consumption } = useSelector((state: RootState) => {
     return state.workflowLoader;
   });
   const dispatch = useDispatch<AppDispatch>();
@@ -29,8 +29,9 @@ const ContextSettings = () => {
         checked={readOnly}
         onChange={(_, checked) => dispatch(setReadOnly(!!checked))}
       />
-      <Checkbox label="Is Monitoring View" checked={monitoringView} onChange={changeMonitoringView} />
+      <Checkbox label="Monitoring View" checked={monitoringView} onChange={changeMonitoringView} />
       <Checkbox label="Dark Mode" checked={darkMode} onChange={(_, checked) => dispatch(setDarkMode(!!checked))} />
+      <Checkbox label="Consumption" checked={consumption} onChange={(_, checked) => dispatch(setConsumption(!!checked))} />
     </div>
   );
 };