+++ libs/data-mapper/src/lib/components/nodeCard/SchemaCard.tsx
@@ -1,4 +1,4 @@
-import { childTargetNodeCardWidth, schemaNodeCardHeight, schemaNodeCardDefaultWidth } from '../../constants/NodeConstants';
+import { childTargetNodeCardWidth, schemaNodeCardDefaultWidth, schemaNodeCardHeight } from '../../constants/NodeConstants';
 import { ReactFlowNodeType } from '../../constants/ReactFlowConstants';
 import { removeSourceSchemaNodes, setCurrentTargetSchemaNode } from '../../core/state/DataMapSlice';
 import type { AppDispatch, RootState } from '../../core/state/Store';
@@ -11,8 +11,8 @@ import { iconForNormalizedDataType } from '../../utils/Icon.Utils';
 import { isSchemaNodeExtended } from '../../utils/Schema.Utils';
 import { ItemToggledState } from '../tree/TargetSchemaTreeItem';
 import HandleWrapper from './HandleWrapper';
-import { getStylesForSharedState, selectedCardStyles } from './NodeCard';
 import type { CardProps } from './NodeCard';
+import { getStylesForSharedState, selectedCardStyles } from './NodeCard';
 import {
   Badge,
   Button,
@@ -26,22 +26,20 @@ import {
 } from '@fluentui/react-components';
 import {
   bundleIcon,
+  CheckmarkCircle12Filled,
   ChevronRight16Filled,
   ChevronRight16Regular,
-  Important12Filled,
-  CheckmarkCircle12Filled,
-  CircleHalfFill12Regular,
   Circle12Regular,
+  CircleHalfFill12Regular,
+  Important12Filled,
 } from '@fluentui/react-icons';
 import type { MenuItemOption } from '@microsoft/designer-ui';
-import { MenuItemType, useCardContextMenu, CardContextMenu } from '@microsoft/designer-ui';
+import { CardContextMenu, MenuItemType, useCardContextMenu } from '@microsoft/designer-ui';
 import { useMemo, useRef, useState } from 'react';
 import { useIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 import type { NodeProps } from 'reactflow';
 import { Position } from 'reactflow';
-
-const contentBtnWidth = schemaNodeCardDefaultWidth - 30;
 
 const useStyles = makeStyles({
   container: {
@@ -83,7 +81,6 @@ const useStyles = makeStyles({
   },
   contentButton: {
     height: '48px',
-    width: `${contentBtnWidth}px`,
     ...shorthands.border('0px'),
     ...shorthands.padding('0px'),
     marginRight: '0px',
@@ -281,7 +278,7 @@ export const SchemaCard = (props: NodeProps<SchemaCardProps>) => {
 
   const targetCardWidth = isChild ? childTargetNodeCardWidth : schemaNodeCardDefaultWidth;
   const cardWidth = isSourceSchemaNode && schemaNode.width ? schemaNode.width : targetCardWidth;
-  const maxWidthCalculated = maxWidth || 0;
+  const maxWidthCalculated = maxWidth || cardWidth;
   const sourceCardMargin = isSourceSchemaNode ? maxWidthCalculated - cardWidth : 0;
 
   return (
@@ -304,7 +301,13 @@ export const SchemaCard = (props: NodeProps<SchemaCardProps>) => {
         />
         {!isInputValid && <Badge size="small" icon={<ExclamationIcon />} color="danger" className={classes.errorBadge} />}
         {connectionStatusIcon && <span style={{ position: 'absolute', right: -16, top: 0 }}>{connectionStatusIcon}</span>}
-        <Button disabled={!!disabled} onClick={onClick} appearance={'transparent'} className={classes.contentButton}>
+        <Button
+          disabled={!!disabled}
+          onClick={onClick}
+          appearance={'transparent'}
+          className={classes.contentButton}
+          style={{ paddingRight: isSourceSchemaNode ? '10px' : '0px' }}
+        >
           <span className={classes.cardIcon}>
             <BundledTypeIcon />
           </span>
@@ -315,11 +318,7 @@ export const SchemaCard = (props: NodeProps<SchemaCardProps>) => {
             visible={shouldNameTooltipDisplay && isTooltipEnabled}
             onVisibleChange={(_ev, data) => setIsTooltipEnabled(data.visible)}
           >
-            <div
-              ref={schemaNameTextRef}
-              className={classes.cardText}
-              style={{ width: !isSourceSchemaNode ? `${contentBtnWidth}px` : '136px' }}
-            >
+            <div ref={schemaNameTextRef} className={classes.cardText} style={{ width: `${cardWidth - 30}px` }}>
               {schemaNode.name}
             </div>
           </Tooltip>
@@ -335,7 +334,6 @@ export const SchemaCard = (props: NodeProps<SchemaCardProps>) => {
           />
         )}
         {
-          // danielle maybe show blank menu for target node? Kinda odd that the regular right-click loads
           <CardContextMenu
             title={'remove'}
             contextMenuLocation={contextMenu.location}