+++ libs/data-mapper/src/lib/components/nodeCard/functionCard/SimpleFunctionCard.tsx
@@ -2,12 +2,12 @@ import { ReactFlowNodeType } from '../../../constants/ReactFlowConstants';
 import { customTokens } from '../../../core';
 import { deleteCurrentlySelectedItem, setSelectedItem } from '../../../core/state/DataMapSlice';
 import type { RootState } from '../../../core/state/Store';
-import { getIconForFunction } from '../../../utils/Icon.Utils';
+import { FunctionIcon } from '../../functionIcon/FunctionIcon';
 import HandleWrapper from './../HandleWrapper';
-import { getStylesForSharedState, selectedCardStyles } from './../NodeCard';
+import { errorCardStyles, getStylesForSharedState, selectedCardStyles } from './../NodeCard';
 import type { FunctionCardProps } from './FunctionCard';
 import { inputsValid, shouldDisplaySourceHandle, shouldDisplayTargetHandle, useFunctionCardStyles } from './FunctionCard';
-import { Button, mergeClasses, PresenceBadge, Text, Tooltip } from '@fluentui/react-components';
+import { Button, mergeClasses, PresenceBadge, Text, tokens, Tooltip } from '@fluentui/react-components';
 import type { MenuItemOption } from '@microsoft/designer-ui';
 import { CardContextMenu, MenuItemType, useCardContextMenu } from '@microsoft/designer-ui';
 import { useMemo, useState } from 'react';
@@ -19,6 +19,7 @@ import { Position } from 'reactflow';
 export const SimpleFunctionCard = (props: NodeProps<FunctionCardProps>) => {
   const { functionData, disabled, functionBranding, displayHandle, onClick, dataTestId } = props.data;
   const reactFlowId = props.id;
+
   const dispatch = useDispatch();
   const classes = useFunctionCardStyles();
   const mergedClasses = mergeClasses(getStylesForSharedState().root, classes.root);
@@ -65,6 +66,13 @@ export const SimpleFunctionCard = (props: NodeProps<FunctionCardProps>) => {
     return inputsValid(reactFlowId, functionData, connections);
   }, [connections, reactFlowId, functionData]);
 
+  const cardStyle =
+    isCurrentNodeSelected || sourceNodeConnectionBeingDrawnFromId === reactFlowId
+      ? { ...selectedCardStyles, backgroundColor: customTokens[functionBranding.colorTokenName] }
+      : areCurrentInputsValid
+      ? { backgroundColor: customTokens[functionBranding.colorTokenName] }
+      : { ...errorCardStyles, backgroundColor: customTokens[functionBranding.colorTokenName] };
+
   return (
     <div
       onContextMenu={contextMenu.handle}
@@ -81,7 +89,18 @@ export const SimpleFunctionCard = (props: NodeProps<FunctionCardProps>) => {
         nodeReactFlowId={reactFlowId}
       />
 
-      {!areCurrentInputsValid && <PresenceBadge size="extra-small" status="busy" className={classes.errorBadge} />}
+      {!areCurrentInputsValid && (
+        <PresenceBadge
+          size="extra-small"
+          status="busy"
+          style={{
+            position: 'absolute',
+            top: '0px',
+            right: '0px',
+            zIndex: 1,
+          }}
+        />
+      )}
 
       <Tooltip
         content={{
@@ -89,17 +108,13 @@ export const SimpleFunctionCard = (props: NodeProps<FunctionCardProps>) => {
         }}
         relationship="label"
       >
-        <Button
-          onClick={onClick}
-          className={mergedClasses}
-          style={
-            isCurrentNodeSelected || sourceNodeConnectionBeingDrawnFromId === reactFlowId
-              ? { ...selectedCardStyles, backgroundColor: customTokens[functionBranding.colorTokenName] }
-              : { backgroundColor: customTokens[functionBranding.colorTokenName] }
-          }
-          disabled={!!disabled}
-        >
-          {getIconForFunction(functionData.functionName, functionData.category, functionData.iconFileName, functionBranding)}
+        <Button onClick={onClick} className={mergedClasses} style={cardStyle} disabled={!!disabled}>
+          <FunctionIcon
+            name={functionData.functionName}
+            categoryName={functionData.category}
+            fileName={functionData.iconFileName}
+            color={tokens.colorNeutralForegroundInverted}
+          />
         </Button>
       </Tooltip>